requirement.txt
bs4
flask
request
jason
response
boto3
re